# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FitHub API",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/injuries": {
      "get": {
        "tags": [
          "injuries"
        ],
        "summary": "List Injuries",
        "operationId": "list_injuries_api_v1_injuries_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/InjuryOut"
                  },
                  "type": "array",
                  "title": "Response List Injuries Api V1 Injuries Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "injuries"
        ],
        "summary": "Report Injury",
        "operationId": "report_injury_api_v1_injuries_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportInjuryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InjuryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/injuries/{injury_id}/status": {
      "patch": {
        "tags": [
          "injuries"
        ],
        "summary": "Update Injury Status",
        "operationId": "update_injury_status_api_v1_injuries__injury_id__status_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "injury_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Injury Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInjuryStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InjuryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/notifications": {
      "get": {
        "tags": [
          "notifications"
        ],
        "summary": "List Notifications",
        "operationId": "list_notifications_api_v1_notifications_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "include_read",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Read"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Notification"
                  },
                  "title": "Response List Notifications Api V1 Notifications Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notifications/{notification_id}/read": {
      "post": {
        "tags": [
          "notifications"
        ],
        "summary": "Mark Read",
        "operationId": "mark_read_api_v1_notifications__notification_id__read_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "notification_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Notification Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Notification"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/training-partners": {
      "get": {
        "tags": [
          "notifications"
        ],
        "summary": "List Training Partners",
        "operationId": "list_training_partners_api_v1_training_partners_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TrainingPartner"
                  },
                  "type": "array",
                  "title": "Response List Training Partners Api V1 Training Partners Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "notifications"
        ],
        "summary": "Add Training Partner",
        "operationId": "add_training_partner_api_v1_training_partners_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddPartnerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingPartner"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddPartnerRequest": {
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 254,
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "AddPartnerRequest"
      },
...
      "BodyRegion": {
        "type": "string",
        "enum": [
          "shoulder",
          "knee",
          "hip",
          "lower_back",
          "wrist",
          "elbow",
          "ankle",
          "neck",
          "hamstring",
          "quad",
          "calf",
          "glute",
          "upper_back",
          "chest",
          "bicep",
          "tricep",
          "lat",
          "hip_flexor",
          "it_band",
          "forearm",
          "rotator_cuff",
          "patellar_tendon",
          "lateral_elbow",
          "medial_elbow",
          "arch",
          "achilles",
          "shin",
          "groin",
          "si_joint",
          "other"
        ],
        "title": "BodyRegion"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InjuryOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "body_region": {
            "$ref": "#/components/schemas/BodyRegion"
          },
          "pain_level": {
            "type": "integer",
            "title": "Pain Level"
          },
          "mechanism": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mechanism"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "cleared_with_restrictions",
              "permanent",
              "resolved"
            ],
            "title": "Status",
            "default": "active"
          },
          "requires_referral": {
            "type": "boolean",
            "title": "Requires Referral"
          },
          "substitutions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Substitutions",
            "default": []
          },
          "contraindicated": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Contraindicated",
            "default": []
          },
          "reported_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reported At"
          },
          "resolved_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "cleared_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cleared At"
          },
          "restriction_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Restriction Notes"
          },
          "staleness_days": {
            "type": "integer",
            "title": "Staleness Days",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "body_region",
          "pain_level",
          "active",
          "requires_referral",
          "staleness_days"
        ],
        "title": "InjuryOut"
      },
...
      "Notification": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "type": {
            "$ref": "#/components/schemas/NotificationType"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "read_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Read At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "type",
          "payload",
          "read_at",
          "created_at"
        ],
        "title": "Notification"
      },
      "NotificationType": {
        "type": "string",
        "enum": [
          "team_session_linked",
          "team_session_updated",
          "workout_link_pending"
        ],
        "title": "NotificationType"
      },
...
      "ReportInjuryRequest": {
        "properties": {
          "body_region": {
            "$ref": "#/components/schemas/BodyRegion"
          },
          "pain_level": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 0.0,
            "title": "Pain Level"
          },
          "mechanism": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "overuse",
                  "acute",
                  "unknown"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Mechanism"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "body_region",
          "pain_level"
        ],
        "title": "ReportInjuryRequest"
      },
...
      "TrainingPartner": {
        "properties": {
          "user_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "guest_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Guest Name"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "session_count": {
            "type": "integer",
            "title": "Session Count"
          },
          "most_common_format": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Most Common Format"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "guest_name",
          "display_name",
          "session_count",
          "most_common_format"
        ],
        "title": "TrainingPartner"
      },
      "UpdateInjuryStatusRequest": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "cleared_with_restrictions",
              "resolved",
              "permanent"
            ],
            "title": "Status"
          },
          "restriction_notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Restriction Notes"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "UpdateInjuryStatusRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}